added roc curves

+++ test_flow.py
@@ -19,15 +19,12 @@ from align_faces import extract_faces, align_faces
 from dataset import FaceDataset
 from openface import load_openface, preprocess_batch
 from classifiers.binary_face_classifier import BinaryFaceClassifier, BinaryFaceNetwork
-from numpy import save, load
+# from numpy import save, loadsa
 
 CONF_THRESHOLD = 0.6
 CONF_TO_STORE = 30
 
-#TODO: jitter training data with gaussian noise and saturation.
-#TODO: SVM running on unknown class
-
-def capture_faces(seconds=10, sampling_duration=0.1, debug=False, imgs = ""):
+def capture_faces(seconds=3, sampling_duration=0.08, debug=False):
     print("Capturing! about to capture {} seconds of video".format(seconds))
     start_time = time.time()
 
@@ -35,10 +32,12 @@ def capture_faces(seconds=10, sampling_duration=0.1, debug=False, imgs = ""):
     face_locs = []
     # frames stores the actual images
     frames = []
+    copyimgs = train_imgs
+    random.shuffle(copyimgs)
     ctr = 1
     while time.time() - start_time < seconds:
         # ret, frame = video_capture.read()
-        for img in imgs[:200]:
+        for img in copyimgs:
             frame = cv2.imread(img)
             faces = extract_faces(frame)
             if len(faces) == 1:
@@ -90,25 +89,25 @@ def retrain_classifier(clf):
     return clf, idx_to_name
 
 
-def add_face(clf, num_classes, imgs):
+def add_face(clf, num_classes, imgs, test_flag):
     name = testname
     if not args["train"]:
         name = input("We don't recognize you! Please enter your name:\n").strip().lower()
     increment = 1
     live_embeddings_loc = "data/embeddings/live"
-    # while name in name_to_idx:
-    #     print("Face exists, append to embeddings!")
-    #     existing_face = np.load("data/embeddings/{}.npy".format(name))
-    #     with open('myfile.npy', 'ab') as f_handle:
-    #         np.save(f_handle, Matrix)
-    #     np.save("data/embeddings/{}.npy".format(name), embeddings)
-    #     return retrain_classifier(clf), 0
-    if name == "skip" or args["test"]:
+    while name in name_to_idx:
+        print("Face exists, append to embeddings!")
+        existing_face = np.load("data/embeddings/{}.npy".format(name))
+        with open('myfile.npy', 'ab') as f_handle:
+            np.save(f_handle, Matrix)
+        np.save("data/embeddings/{}.npy".format(name), embeddings)
         return retrain_classifier(clf), 0
-    samples = capture_faces(imgs= imgs)
+    if name == "skip" or test_flag:
+        return retrain_classifier(clf), 0
+    samples = capture_faces()
     while len(samples) < 50:
         print("We could not capture sufficient samples. Please try again.\n")
-        samples = capture_faces(imgs =imgs)
+        samples = capture_faces()
     embeddings = preprocess_batch(samples)
     embeddings = openFace(embeddings)
     embeddings = embeddings.detach().numpy()
@@ -118,6 +117,9 @@ def add_face(clf, num_classes, imgs):
         embeddings = update_embedding(live_embeddings_loc, embeddings, name)
         increment = 0
     # save name and embeddings
+    print("embeddings should be saved now")
+    print(name)
+    print(live_embeddings_loc + "/{}.npy".format(name))
     np.save(live_embeddings_loc + "/{}.npy".format(name), embeddings)
     return retrain_classifier(clf), increment
 
@@ -141,25 +143,21 @@ def load_model():
     clf = clf.fit(data, labels)
     return clf, num_classes, ds.ix_to_name
 
-def recognize(clf, num_classes, idx_to_name, imgs):
+def recognize(clf, num_classes, idx_to_name, imgs, test_flag=False):
     # to store previous confidences to determine whether a face exists
     prev_conf = deque(maxlen=CONF_TO_STORE)
     f_count = 1
     global CONF_THRESHOLD 
     names = [] 
     print("Starting the test...")
-    # video_capture = cv2.VideoCapture("test_faces/test_videos/lily_singh1.mov")
     # frame_array = []
     test_results = [] 
     CONF_THRESHOLD += 0.2/num_classes
     print(CONF_THRESHOLD)
     for image in imgs:
-        print(image)
-        # ret is error code but we don't care about it
         # ret, frame = video_capture.read()
         frame = cv2.imread(image) 
         rects = extract_faces(frame)
-        # print(frame)
         if len(rects) > 0:
 
             # draw all bounding boxes for faces
@@ -182,37 +180,36 @@ def recognize(clf, num_classes, idx_to_name, imgs):
             names = [idx_to_name[idx] for idx in predictions]
             # replace all faces below confidence w unknown
             names = [names[i] if probs[i] > CONF_THRESHOLD else "unknown_class" for i in range(len(probs))]
-            # print("Hi {}!".format(names))
+            print("Hi {}!".format(names))
             for i in range(len(names)):
                 x, y, w, h = face_utils.rect_to_bb(rects[i])
                 cv2.putText(frame, names[i], (x, y), cv2.FONT_HERSHEY_PLAIN, 1.5, (0, 255, 0), 2)
-                test_results.append(names[i]) 
+            test_results.append(names[0]) 
             # determine if we need to trigger retraining
             # we only retrain if there is one person in the frame and they are unrecognized or there are 0 classes
             if len(faces) == 1:
                 prev_conf.append(unknown_class_prob)
-                if np.mean(prev_conf) > CONF_THRESHOLD and len(prev_conf) == CONF_TO_STORE:
-                    # (clf, idx_to_name), inc = add_face(clf, num_classes, imgs)
+                if np.mean(prev_conf) > CONF_THRESHOLD and len(prev_conf) == CONF_TO_STORE and (test_flag==False):
+                    (clf, idx_to_name), inc = add_face(clf, num_classes, imgs, test_flag)
                     print(idx_to_name)
                     inc = 0
                     num_classes += inc
                     prev_conf.clear()
         else:
             print("No faces detected.")
+            test_results.append("no detection") 
         cv2.imshow('Camera Feed', frame)
-        if cv2.waitKey(1) & 0xFF == ord('r'):
-            # (clf, idx_to_name), inc = add_face(clf, num_classes, imgs)
+        if cv2.waitKey(1) & 0xFF == ord('r') and (test_flag==False):
+            (clf, idx_to_name), inc = add_face(clf, num_classes, imgs, test_flag)
             print(idx_to_name)
             num_classes += inc
             prev_conf.clear()
         # else: 
             # print("ERROR: no frame captured")
-    # tmod.plot_acc(testname, test_results)
-    # save("data/test/test_results/accs/" + testname + ".txt", test_results) 
-    # write_log(test_results, testname.split(".")[0])
+            # test_results.append("no detection") 
     all_results.append(test_results)
-    print(all_results)
-    # print(test_results)
+    return all_results
+
     # video_capture.release()
     cv2.destroyAllWindows()
 
@@ -249,7 +246,6 @@ if __name__ == "__main__":
     print("Using device {}".format(device))
     openFace = load_openface(device) 
 
-
     all_results = []
     # KEEP
     # if args["test"]:
@@ -269,8 +265,9 @@ if __name__ == "__main__":
     if args["clean"]:
         files = glob.glob('data/embeddings/live/*.npy')
         print(files)
-        for f in files[:-1]:
-            os.remove(f)
+        if len(files) != 1:
+            for f in files[:-1]:
+                os.remove(f)
         print(glob.glob('data/embeddings/live/*.npy'))
 
     if args["train"] or  args["test"]:
@@ -279,26 +276,30 @@ if __name__ == "__main__":
         test_imgs = []
 
         # Choose one random train person
-        rand = files[random.randint(0,500)]
+        rand = files[random.randint(0,100)]
         train_imgs = glob.glob(rand+ "/*")[:100]
         print("chosen rand was: " + rand)
-        for folder in files[:90]:
+        for folder in files[100:104]:
             imgs = glob.glob(folder+ "/*")
             # if len (imgs)>300:
             #     train_imgs.append(imgs[:200]) 
             #     test_imgs.append(imgs[200:300]) 
-            # train_imgs.append(imgs[:200]) 
-            # test_imgs.append(imgs[200:300])
-            test_imgs.append(imgs[:50]) 
-        test_imgs.append(glob.glob("data/vgff2_test/n001838"+ "/*")[50:])
-        # test_imgs = test_imgs.flatten()
-        # print(test_imgs)
+            # train_imgs.append(imgs[:100])
+            test_imgs += imgs[:100]
+        # train_imgs = train_imgs[:10] 
+
+        # Add the train person to the test data
+        train_remains = glob.glob(rand+ "/*")[100:200]
+        test_imgs += train_remains
+
+        random.shuffle(test_imgs)
+        np.save("data/test/truth_labels.npy", np.array(test_imgs))
+        np.save("data/test/train_labels.npy", np.asarray(train_imgs))
+
+        print("LENGTHS")
         print(len(test_imgs)) 
+        print(len(train_imgs)) 
         start = 0
-        # trains = [x for x in files if "train" in x ]
-        # devs = [x for x in files if "dev" in x ]
-        # tests = [x for x in files if "testing" in x ]
-        # files = []
         mode = "Test case one vs unknowns"
         # if args["train"]:
         #     files += trains + devs
@@ -307,39 +308,62 @@ if __name__ == "__main__":
         #     files += tests
         #     mode += "Testing "
         # print(files)
-        random.shuffle(test_imgs)
-        for i in tqdm(range(start, len(test_imgs)), total=len(test_imgs)):
-            file = test_imgs[i]
-            # video_capture = cv2.VideoCapture(file)
-            testname = "test all"
-            print(mode + " for file " + testname)
-            name_to_idx = {idx_to_name[idx]: idx for idx in idx_to_name}
-            recognize(clf, num_classes, idx_to_name, file)
-            clf, num_classes, idx_to_name = load_model()
-            print("DONE!! Mode was " + mode)
-        # tmod.plot()
 
-    print("Writing test result logs")
-    file1 = open("data/test/test_results/" +  "ALLtest_resultss.txt", "a+")  # append mode
-    file1.write('Test results for ' + "data/vgff2_test/n001838.npy"+ datetime.datetime.now().strftime(" on %Y-%m-%d %H:%M:%S") +"\n")
-    # print('\n'.join(test_res))
-    # file1.write('\n'.join(test_res))
-    all_results = flatten(np.asarray(all_results))
-    file1.write("Accuracy: \n")
-    for name in set(all_results):
-        file1.write("   "+name + " {}%\n".format(all_results.count(name)/len(all_results)*100))
- 
-    save("data/test/test_results/accs/" + all_test + ".npy", np.asarray(all_results)) 
-    
+        # TEST MODE for VIDEOS
+        # dataset_mode = train_imgs
+        # if args['test']: dataset_mode = test_imgs
+        # for i in tqdm(range(start, len(dataset_mode)), total=len(dataset_mode)):
+        #     file = dataset_mode[i]
+        #     # video_capture = cv2.VideoCapture(file)
+        #     mode = "TRAIN"
+        #     print(file[0])
+        #     testname = file[0].split("/")[-2]
+        name_to_idx = {idx_to_name[idx]: idx for idx in idx_to_name}
+        #     recognize(clf, num_classes, idx_to_name, file)
+        #     clf, num_classes, idx_to_name = load_model()
+        #     print("DONE!! Mode was " + mode)
+        # tmod.plot()
+        if args["train"]:
+            print("NOW STARTING TRAIN MODE")
+            testname = train_imgs[0].split("/")[-2]
+            train_results = recognize(clf, num_classes, idx_to_name, train_imgs, test_flag = False)
+            print(train_results)
+            clf, num_classes, idx_to_name = load_model()
+            print("TRAINING DONE")
+
+        if args["test"]:
+            print("NOW STARTING TEST MODE")
+            test_imgs = np.asarray(test_imgs).flatten()
+            all_test_results = recognize(clf, num_classes, idx_to_name, test_imgs, test_flag = True)
+            print(all_test_results)
+            clf, num_classes, idx_to_name = load_model()
+            print("TESTING DONE")
+
+
+    # print("Writing test result logs")
+    # file1 = open("data/test/test_results/" +  "ALLtest_resultss.txt", "a+")  # append mode
+    # file1.write('Test results for ' + "all vs mixed"+ datetime.datetime.now().strftime(" on %Y-%m-%d %H:%M:%S") +"\n")
+    # # print('\n'.join(test_res))
+    # # file1.write('\n'.join(test_res))
+    # all_results = np.asarray(all_results).flatten()
+    # np.save("data/test/test_results.npy", test_results)
+    # np.save("data/test/test_results/accs/" + "all_test" + ".npy", np.asarray(all_results)) 
+    # # all_results = list(all_results)
+    # file1.write("Accuracy: \n")
+    # for name in set(all_results):
+    #     file1.write("   "+name + " {}%\n".format(all_results.count(name)/len(all_results)*100))
+
+    np.save("data/test/test_results/accs/" + "all_test" + ".npy", np.asarray(all_test_results)) 
+    print(len(test_imgs))
+    print(len(all_test_results))
+    print(len(all_test_results[0]))
+    tmod.accuracy_metrics(test_imgs, all_test_results)
+    print("ALL DONE!!")
 
     if args["live"]:
         print("Starting video capture...")
         video_capture = cv2.VideoCapture(0)
         recognize(clf, num_classes, idx_to_name, args["test"])
-
-
-
-
 
     # clf, num_classes, idx_to_name = load_model()
     # print(idx_to_name)
